#25 restructured remaining mocks and created a data service to retrieve the mocks
