--- src/app/_services/appointment.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AppointmentService } from './appointment.service';
-
-describe('AppointmentService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: AppointmentService = TestBed.get(AppointmentService);
-    expect(service).toBeTruthy();
-  });
-});
